pyproject.toml
[project]
name = "mullardoch"
version = "0.1.0"
description = "Python modules for the UGV02 robot"
authors = [
    {name = "a.b.christie",email = "alan.christie@matildapeak.com"}
]
license = {text = "Apache License 2.0"}
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "apscheduler (>=3.11.2,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13,<4.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
